Update chore status

+++ buupass/apps/chores/views.py
@@ -1,3 +1,4 @@
+from datetime import datetime
 from django.http import Http404
 from rest_framework.views import APIView
 from rest_framework.response import Response
@@ -8,14 +9,15 @@ from rest_framework.permissions import IsAuthenticated
 from buupass.helpers.endpoint_response import \
     get_success_responses
 from .models import Chores
+from buupass.apps.authentication.models import \
+    User
 from .serializers import ChoresSerializer
 
 
 class ChoresList(APIView):
-    """
-    List all chores, or create a new chore.
-    """
     permission_classes = (IsAuthenticated,)
+
+    """List all chores"""
 
     def get(self, request, format=None):
         chores = Chores.objects.all()
@@ -26,20 +28,31 @@ class ChoresList(APIView):
             status_code=status.HTTP_201_CREATED
         )
 
+    """Create new chores"""
+
     def post(self, request, format=None):
-        serializer = ChoresSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return get_success_responses(
-                data=serializer.data,
-                message="Successfully created a chore",
-                status_code=status.HTTP_201_CREATED
-            )
-        return get_success_responses(
-                data=serializer.errors,
-                message="Error Message",
-                status_code=status.HTTP_400_BAD_REQUEST
-            )
+        try:
+            user_id = request.user.id
+            user = User.objects.get(id=user_id)
+
+            if (user.type == "OWNER"):
+                serializer = ChoresSerializer(data=request.data)
+                if serializer.is_valid():
+                    serializer.save()
+                    return get_success_responses(
+                        data=serializer.data,
+                        message="Successfully created a chore",
+                        status_code=status.HTTP_201_CREATED
+                    )
+                return get_success_responses(
+                    data=serializer.errors,
+                    message="Error Message",
+                    status_code=status.HTTP_400_BAD_REQUEST
+                )
+            return Response("You are not allowed to perform this action")
+        except Exception as e:
+            return Response({"Error message": "{}".format(e)})
+
 
 class ChoresDetail(APIView):
     """Get a specific chore"""
@@ -53,19 +66,27 @@ class ChoresDetail(APIView):
 
     def put(self, request, pk, format=None):
         chore = self.get_object(pk)
-        serializer = ChoresSerializer(chore, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return get_success_responses(
-                data=serializer.data,
-                message="Successfully created a chore",
-                status_code=status.HTTP_200_OK
-            )
-        return get_success_responses(
-                data=serializer.errors,
-                message="Error Message",
-                status_code=status.HTTP_400_BAD_REQUEST
-            )
+        try:
+            user_id = request.user.id
+            user = User.objects.get(id=user_id)
+
+            if (user.type == "OWNER"):
+                serializer = ChoresSerializer(chore, data=request.data)
+                if serializer.is_valid():
+                    serializer.save()
+                    return get_success_responses(
+                        data=serializer.data,
+                        message="Successfully updated the chore",
+                        status_code=status.HTTP_200_OK
+                    )
+                return get_success_responses(
+                    data=serializer.errors,
+                    message="Error Message",
+                    status_code=status.HTTP_400_BAD_REQUEST
+                )
+            return Response("You are not allowed to perform this action")
+        except Exception as e:
+            return Response({"Error message": "{}".format(e)})
 
     def get(self, request, pk, format=None):
         chore = self.get_object(pk)
@@ -75,9 +96,30 @@ class ChoresDetail(APIView):
             message="Successfully fetched the chore",
             status_code=status.HTTP_200_OK
         )
-   
 
     def delete(self, request, pk, format=None):
         chore = self.get_object(pk)
-        chore.delete()
-        return Response({"Success: uccessfully deleted the chore"}, status=status.HTTP_200_OK)
+        try:
+            user_id = request.user.id
+            user = User.objects.get(id=user_id)
+
+            if (user.type == "OWNER"):
+                chore.delete()
+                return Response({"Success: successfully deleted the chore"}, status=status.HTTP_200_OK)
+            return Response("You are not allowed to perform this action")
+        except Exception as e:
+            return Response({"Error message": "{}".format(e)})
+
+
+class ChoreUpdate(APIView):
+    """Update status of the chore"""
+
+    def put(self, request, pk, format=None):
+        chore = Chores.objects.get(id=pk)
+        serializer = ChoresSerializer(chore, data=request.data)
+        chore.completed = True
+        chore.save()
+
+        return Response(
+            "Chore completed successfully"
+        )
